% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_stack_weights.r
\name{apply_stack_weights}
\alias{apply_stack_weights}
\title{Applies stack weights}
\usage{
apply_stack_weights(ebma_out, stack_out, L2.unit, L1.x, preds_all, y)
}
\arguments{
\item{L2.unit}{Geographic unit. A character scalar containing the column
name of the geographic unit in \code{survey} and \code{census} at which
outcomes should be aggregated.}

\item{L1.x}{Individual-level covariates. A character vector containing the
column names of the individual-level variables in \code{survey} and
\code{census} used to predict outcome \code{y}. Note that geographic unit
is specified in argument \code{L2.unit}.}

\item{y}{Outcome variable. A character scalar containing the column name of
the outcome variable in \code{survey}.}
}
\description{
\code{apply_stack_weights} applies stacking weights to the census level
predictions.
}
