% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_data.R
\docType{data}
\name{census}
\alias{census}
\title{A file that mimics the national census.}
\format{
A data frame with 2934 rows and 13 variables:
\describe{
  \item{state}{The state that the respondent is from}
  \item{L2.unit}{An aditional  variable indicating the state that the respondent is from}
  \item{region}{Region in the U.S. that the respondent is from; 1 = something, 2 = something, 3 = something, 4 = something}
  \item{L1x1}{Age group (four categories)}
  \item{L1x2}{Education level (four categories)}
  \item{L1x3}{Gender-race combination (six categories)}
  \item{proportion}{The proportion of respondents of that ideal type in the population}
  \item{L2.x1}{The share of votes for the Republican candidate in the previous presidential election}
  \item{L2.x2}{The percentage of Evangelical Protestant or Mormon respondents}
  \item{L2.x3}{The percentage of the population living in urban areas}
  \item{L2.x4}{The unemployment rate}
  \item{L2.x5}{The share of Hispanics}
  \item{L2.x6}{The share of Whites}
  ...
}
}
\source{
\url{https://cces.gov.harvard.edu/pages/welcome-cooperative-congressional-election-study}
}
\usage{
data(census)

It is generated from the full 2008 Cooperative Congressional Election Studies item cc418_1 by dissaggregating the 64 ideal type combinations of the individual level variables L1x1, L2x2 and L1x3. A row is an ideal type in a given state.
}
\description{
A file that mimics the national census.
}
\keyword{datasets}
