% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mrp_classifier.r
\name{deep_mrp_classifier}
\alias{deep_mrp_classifier}
\title{Deep MrP classifier}
\usage{
deep_mrp_classifier(y, form, data, verbose)
}
\arguments{
\item{y}{Outcome variable. A character vector containing the column names of
the outcome variable. A character scalar containing the column name of
the outcome variable in \code{survey}.}

\item{form}{Model formula. A two-sided linear formula describing
the model to be fit, with the outcome on the LHS and the covariates
separated by + operators on the RHS.}

\item{data}{Data. A data.frame containing the data used to train the model.}

\item{verbose}{Verbose output. A logical argument indicating whether or not
verbose output should be printed. Default is \code{FALSE}.}
}
\value{
A Deep MrP model. A \code{\link[vglmer]{vglmer}} object.
}
\description{
\code{deep_mrp_classifier} applies Deep MrP implemented in the \pkg{vglmer}
package to a data set.
}
