% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gb_classifier.R
\name{gb_classifier_update}
\alias{gb_classifier_update}
\title{GB classifier update}
\usage{
gb_classifier_update(object, n.new.trees, verbose = c(TRUE, FALSE))
}
\arguments{
\item{object}{Gradient tree boosting output. A gbm object.}

\item{n.new.trees}{Number of additional trees to grow. A numeric scalar.}

\item{verbose}{Verbose output. A logical vector indicating whether or not
verbose output should be printed.}
}
\value{
An updated gradient tree boosting model. A \code{\link[gbm]{gbm.more}} object.
}
\description{
\code{gb_classifier_update()} grows additional trees in gradient tree
boosting ensemble.
}
