% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_classifier.R
\name{knn_classifier}
\alias{knn_classifier}
\title{KNN classifier}
\usage{
knn_classifier(
  y,
  form,
  data.train,
  data.valid,
  knn.k.value,
  knn.kernel = c("rectangular", "triangular", "epanechnikov", "biweight", "triweight",
    "cos", "inv", "gaussian", "optimal"),
  verbose = c(TRUE, FALSE),
  ...
)
}
\arguments{
\item{y}{Outcome variable. A character scalar containing the column name of
the outcome variable in \code{survey}.}

\item{form}{Model formula. A two-sided linear formula describing
the model to be fit, with the outcome on the LHS and the covariates
separated by + operators on the RHS.}

\item{data.train}{Training data. A \code{data.frame} containing the
cross-validation data used to train the model.}

\item{data.valid}{Validation data. A \code{data.frame} containing the
cross-validation data used to evaluate the model.}

\item{knn.k.value}{KNN number of neighbors. A positive integer-valued scalar
specifying the number of neighbors to be considered in the KNN model.}

\item{knn.kernel}{KNN kernel. A character-valued scalar specifying the kernel
to be used in the KNN model. The possible values are \code{rectangular}
(which is standard unweighted KNN), \code{triangular}, \code{epanechnikov}
(or beta(2,2)), \code{biweight} (or beta(3,3)), \code{triweight} (or
beta(4,4)), \code{cos}, \code{inv}, \code{gaussian}, and \code{optimal}.}

\item{verbose}{Verbose output. A logical vector indicating whether or not
verbose output should be printed.}
}
\value{
The k-nearest neighbors model. A \code{\link[kknn]{kknn}} object.
}
\description{
\code{knn_classifier} applies k-nearest neighbors classification to a data
set.
}
