% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loss_function}
\alias{loss_function}
\title{Estimates loss value}
\usage{
loss_function(
  pred,
  data.valid,
  loss.unit = c("individuals", "L2 units"),
  loss.fun = c("MSE", "MAE"),
  y,
  L2.unit
)
}
\description{
\code{loss_function()} estimates the loss based on a loss function.
}
