% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multicore}
\alias{multicore}
\title{Register cores for multicore computing}
\usage{
multicore(cores = 1, type, cl = NULL)
}
\arguments{
\item{cores}{Number of cores to be used. An integer. Default is \code{1}.}

\item{type}{Whether to start or end parallel processing. A character string.
The possible values are \code{open}, \code{close}.}

\item{cl}{The registered cluster. Default is \code{NULL}}
}
\value{
No return value, called to register or un-register clusters for
  parallel processing.
}
\description{
\code{multicore()} registers cores for parallel processing.
}
