% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_stratification.R
\name{post_stratification}
\alias{post_stratification}
\title{Apply post-stratification to classifiers.}
\usage{
post_stratification(
  y,
  L1.x,
  L2.x,
  L2.unit,
  L2.reg,
  best.subset.opt,
  lasso.opt,
  pca.opt,
  gb.opt,
  svm.opt,
  mrp.include,
  n.minobsinnode,
  L2.unit.include,
  L2.reg.include,
  kernel,
  mrp.L2.x,
  data,
  ebma.fold,
  census,
  verbose
)
}
\description{
Apply post-stratification to classifiers.
}
