% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quiet}
\alias{quiet}
\title{Suppress cat in external package}
\source{
The author is Hadley Wickham. We found the function here:
\url{https://r.789695.n4.nabble.com/Suppressing-output-e-g-from-cat-td859876.html}.
}
\usage{
quiet(x)
}
\arguments{
\item{x}{Input. It can be any kind.}
}
\description{
\code{quiet()} suppresses cat output.
}
