% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_best_subset.R
\name{run_best_subset}
\alias{run_best_subset}
\title{Apply best subset classifier to MrP.}
\usage{
run_best_subset(
  y,
  L1.x,
  L2.x,
  L2.unit,
  L2.reg,
  loss.unit,
  loss.fun,
  data,
  verbose,
  cores
)
}
\arguments{
\item{data}{Data for cross-validation. A \code{list} of \eqn{k}
\code{data.frames}, one for each fold to be used in \eqn{k}-fold
cross-validation.}
}
\value{
A model formula of the winning best subset classifier model.
}
\description{
\code{run_best_subset} is a wrapper function that applies the best subset
classifier to a list of models provided by the user, evaluates the models'
prediction performance, and chooses the best-performing model.
}
