% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lasso.R
\name{run_lasso_mc_kfolds}
\alias{run_lasso_mc_kfolds}
\title{Lasso multicore tuning.}
\usage{
run_lasso_mc_kfolds(
  y,
  L1.x,
  L2.x,
  L2.unit,
  L2.reg,
  loss.unit,
  loss.fun,
  data,
  cores,
  loss.function,
  verbose,
  L2.fe.form,
  L1.re,
  lambda.value,
  lasso.classifier
)
}
\arguments{
\item{y}{Outcome variable. A character vector containing the column names of
the outcome variable. A character scalar containing the column name of
the outcome variable in \code{survey}.}

\item{L1.x}{Individual-level covariates. A character vector containing the
column names of the individual-level variables in \code{survey} and
\code{census} used to predict outcome \code{y}. Note that geographic unit
is specified in argument \code{L2.unit}.}

\item{L2.x}{Context-level covariates. A character vector containing the
column names of the context-level variables in \code{survey} and
\code{census} used to predict outcome \code{y}.}

\item{L2.unit}{Geographic unit. A character scalar containing the column
name of the geographic unit in \code{survey} and \code{census} at which
outcomes should be aggregated.}

\item{L2.reg}{Geographic region. A character scalar containing the column
name of the geographic region in \code{survey} and \code{census} by which
geographic units are grouped (\code{L2.unit} must be nested within
\code{L2.reg}). Default is \code{NULL}.}

\item{loss.unit}{Loss function unit. A character-valued scalar indicating
whether performance loss should be evaluated at the level of individual
respondents (\code{individuals}) or geographic units (\code{L2 units}).
Default is \code{individuals}.}

\item{loss.fun}{Loss function. A character-valued scalar indicating whether
prediction loss should be measured by the mean squared error (\code{MSE})
or the mean absolute error (\code{MAE}). Default is \code{MSE}.}

\item{data}{Data for cross-validation. A \code{list} of \eqn{k}
\code{data.frames}, one for each fold to be used in \eqn{k}-fold
cross-validation.}

\item{cores}{The number of cores to be used. An integer indicating the number
of processor cores used for parallel computing. Default is 1.}

\item{loss.function}{Loss function. A character-valued scalar indicating
whether prediction loss should be measured by the mean squared error
(\code{MSE}) or the mean absolute error (\code{MAE}). The value is
inherited from \code{auto_MrP}.}

\item{verbose}{Verbose output. A logical argument indicating whether or not
verbose output should be printed. Default is \code{FALSE}.}

\item{L2.fe.form}{The fixed effects part of the Lasso classifier formula. The
formula is inherited from \code{run_lasso}.}

\item{L1.re}{A list of random effects for the Lasso classifier formula. The
formula is inherited from \code{run_lasso}.}

\item{lambda.value}{The lasso lambda value. A numeric scalar inherited from
\code{run_lasso}.}

\item{lasso.classifier}{Lasso classifier. An \code{autoMrP} function.

#' @return The cross-validation errors for all models. A list.}
}
\description{
\code{run_lasso_mc_kfolds} is called from within \code{run_lasso}. It
tunes using multiple cores.
}
\examples{
\dontrun{
# not yet
}
}
