% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_svm.R
\name{run_svm}
\alias{run_svm}
\title{Apply support vector machine classifier to MrP.}
\usage{
run_svm(
  y,
  L1.x,
  L2.x,
  L2.eval.unit,
  L2.unit,
  L2.reg,
  kernel = "radial",
  loss.fun,
  loss.unit,
  gamma,
  cost,
  data,
  verbose,
  cores
)
}
\arguments{
\item{L2.eval.unit}{Geographic unit. A character scalar containing the column
name of the geographic unit in \code{survey} and \code{census} at which
outcomes should be aggregated.}

\item{kernel}{SVM kernel. A character-valued scalar specifying the kernel to
be used by SVM. The possible values are \code{linear}, \code{polynomial},
\code{radial}, and \code{sigmoid}. Default is \code{radial}.}

\item{gamma}{SVM kernel parameter. A numeric vector whose values specify
the gamma parameter in the SVM kernel. This parameter is needed for all
kernel types except linear. Default is
\eqn{c(0.3, 0.5, 0.55, 0.6, 0.65, 0.7, 0.8, 0.9, 1, 2, 3, 4)}.}

\item{cost}{SVM cost parameter. A numeric vector whose values specify the
cost of constraints violation in SVM. Default is \eqn{c(1, 10)}.}

\item{data}{Data for cross-validation. A \code{list} of \eqn{k}
\code{data.frames}, one for each fold to be used in \eqn{k}-fold
cross-validation.}
}
\value{
The support vector machine tuned parameters. A list.
}
\description{
\code{run_svm} is a wrapper function that applies the support vector machine
classifier to data provided by the user, evaluates prediction performance,
and chooses the best-performing model.
}
