% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearner_predict.r
\name{superlearner_predict}
\alias{superlearner_predict}
\title{Predict individual-level survey respones for new data.}
\usage{
superlearner_predict(obj, data.test, L2.x.scale = TRUE)
}
\arguments{
\item{obj}{An autoMrP object.}

\item{data.test}{A new data frame with the same columns as the original data
used to train autoMrP. Note: This data should not have been used to train
the autoMrP model.}

\item{L2.x.scale}{Scale context-level covariates. A logical argument
indicating whether the context-level covariates should be normalized.
Default is \code{TRUE}. Note that if set to \code{FALSE}, then the
context-level covariates should be normalized prior to calling
\code{auto_MrP()}.}
}
\value{
The individual-level predictions. A tibble. It contains the
individual-level predictions from all estimated classifiers, ebma, and
 stacking.
}
\description{
Predicts individual-level responses for new data based on a trained auto_MrP
object. This is uesful for combining autoMrP predictions with predictions
from other models into a superlearner.
}
