% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_classifier.R
\name{svm_classifier}
\alias{svm_classifier}
\title{SVM classifier}
\usage{
svm_classifier(
  method,
  form,
  data,
  kernel,
  error.fun,
  probability,
  gamma.set,
  cost.set,
  sampling = "cross",
  cross,
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{method}{Function. A character string specifying the name of the
function to be tuned.}

\item{form}{Model formula. A two-sided linear formula describing
the model to be fit, with the outcome on the LHS and the covariates
separated by + operators on the RHS.}

\item{data}{Data. A data.frame containing the cross-validation data used to
train and evaluate the model.}

\item{kernel}{Kernel for SVM. A character string specifying the kernel to
be used for SVM. The possible types are linear, polynomial, radial, and
sigmoid. Default is radial.}

\item{error.fun}{IS_MISSING}

\item{probability}{Probability predictions. A logical argument indicating
whether the model should allow for probability predictions}

\item{gamma.set}{Gamma parameter for SVM. This parameter is needed for all
kernels except linear.}

\item{cost.set}{Cost parameter for SVM. This parameter specifies the cost of
constraints violation.}

\item{sampling}{Sampling scheme. A character string specifying the sampling
scheme to be used. Possible values are cross, fix, and boot. Default is
cross, which performs cross-validation.}

\item{verbose}{Verbose output. A logical vector indicating whether or not
verbose output should be printed.}
}
\value{

}
\description{
\code{svm_classifier} applies support vector machine classification to a
data set.
}
\examples{
#not_yet
}
