% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quiet}
\alias{quiet}
\title{Suppress cat in external package}
\usage{
quiet(x)
}
\arguments{
\item{x}{Input. It can be any kind.}
}
\value{
No return value, called to suppress cat output to the console.
}
\description{
\code{quiet()} suppresses cat output.
}
