% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pca.R
\name{run_pca}
\alias{run_pca}
\title{Apply PCA classifier to MrP.}
\usage{
run_pca(
  y,
  L1.x,
  L2.x,
  L2.unit,
  L2.reg,
  loss.unit,
  loss.fun,
  data,
  cores,
  verbose
)
}
\arguments{
\item{data}{Data for cross-validation. A \code{list} of \eqn{k}
\code{data.frames}, one for each fold to be used in \eqn{k}-fold
cross-validation.}
}
\value{
A model formula of the winning best subset classifier model.
}
\description{
\code{run_pca} is a wrapper function that applies the PCA classifier to data
provided by the user, evaluates prediction performance, and chooses the
best-performing model.
}
