% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_svm.R
\name{run_svm}
\alias{run_svm}
\title{Apply support vector machine classifier to MrP.}
\usage{
run_svm(
  y,
  L1.x,
  L2.x,
  L2.unit,
  L2.reg,
  kernel = "radial",
  loss.fun,
  loss.unit,
  gamma,
  cost,
  data,
  verbose,
  cores
)
}
\arguments{
\item{y}{Outcome variable. A character scalar containing the column name of
the outcome variable in \code{survey}.}

\item{L1.x}{Individual-level covariates. A character vector containing the
column names of the individual-level variables in \code{survey} and
\code{census} used to predict outcome \code{y}. Note that geographic unit
is specified in argument \code{L2.unit}.}

\item{L2.x}{Context-level covariates. A character vector containing the
column names of the context-level variables in \code{survey} and
\code{census} used to predict outcome \code{y}.}

\item{L2.unit}{Geographic unit. A character scalar containing the column
name of the geographic unit in \code{survey} and \code{census} at which
outcomes should be aggregated.}

\item{L2.reg}{Geographic region. A character scalar containing the column
name of the geographic region in \code{survey} and \code{census} by which
geographic units are grouped (\code{L2.unit} must be nested within
\code{L2.reg}). Default is \code{NULL}.}

\item{kernel}{SVM kernel. A character-valued scalar specifying the kernel to
be used by SVM. The possible values are \code{linear}, \code{polynomial},
\code{radial}, and \code{sigmoid}. Default is \code{radial}.}

\item{loss.fun}{SVM loss function. If \code{NULL}, then SVM uses the
misclassification error to measure the loss of categorical predictions and
the mean squared error to measure the loss of numeric predictions. Default
is \code{NULL}.}

\item{loss.unit}{Loss function unit. A character-valued scalar indicating
whether performance loss should be evaluated at the level of individual
respondents (\code{individuals}) or geographic units (\code{L2 units}).
Default is \code{individuals}.}

\item{gamma}{SVM kernel parameter. A numeric vector whose values specify
the gamma parameter in the SVM kernel. This parameter is needed for all
kernel types except linear. Default is
\eqn{c(0.3, 0.5, 0.55, 0.6, 0.65, 0.7, 0.8, 0.9, 1, 2, 3, 4)}.}

\item{cost}{SVM cost parameter. A numeric vector whose values specify the
cost of constraints violation in SVM. Default is \eqn{c(1, 10)}.}

\item{data}{Data for cross-validation. A \code{list} of \eqn{k}
\code{data.frames}, one for each fold to be used in \eqn{k}-fold
cross-validation.}

\item{verbose}{Verbose output. A logical argument indicating whether or not
verbose output should be printed. Default is \code{TRUE}.}

\item{cores}{The number of cores to be used. An integer indicating the number
of processor cores used for parallel computing. Default is 1.}
}
\value{
The support vector machine tuned parameters. A list.
}
\description{
\code{run_svm} is a wrapper function that applies the support vector machine
classifier to data provided by the user, evaluates prediction performance,
and chooses the best-performing model.
}
\examples{
\dontrun{
# Create list of cross-validation folds
cv_folds <- list(
  `1` = survey_item[1:200, ],
  `2` = survey_item[201:400, ],
  `3` = survey_item[401:1500, ])

# Run svm classifier
m <- run_svm(
  y = "YES",
  L1.x = c("L1x1", "L1x2"),
  L2.x = c("L2.x1", "L2.x2"),
  L2.unit = "state",
  L2.reg = "region",
  kernel = "radial",
  loss.fun = "MSE",
  loss.unit = "individuals",
  gamma = c(0.3, 0.1),
  cost = c(1, 50),
  data = cv_folds,
  verbose = TRUE)
}
}
