% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_svm.R
\name{run_svm}
\alias{run_svm}
\title{Apply support vector machine classifier to MrP.}
\usage{
run_svm(
  y,
  L1.x,
  L2.x,
  L2.eval.unit,
  L2.unit,
  L2.reg,
  kernel = "radial",
  loss.fun,
  loss.unit,
  gamma,
  cost,
  data,
  verbose,
  cores
)
}
\arguments{
\item{y}{Outcome variable. A character vector containing the column names of
the outcome variable. A character scalar containing the column name of
the outcome variable in \code{survey}.}

\item{L1.x}{Individual-level covariates. A character vector containing the
column names of the individual-level variables in \code{survey} and
\code{census} used to predict outcome \code{y}. Note that geographic unit
is specified in argument \code{L2.unit}.}

\item{L2.x}{Context-level covariates. A character vector containing the
column names of the context-level variables in \code{survey} and
\code{census} used to predict outcome \code{y}. To exclude context-level
variables, set \code{L2.x = NULL}.}

\item{L2.eval.unit}{Geographic unit for the loss function. A character scalar
containing the column name of the geographic unit in \code{survey} and
\code{census}.}

\item{L2.unit}{Geographic unit. A character scalar containing the column
name of the geographic unit in \code{survey} and \code{census} at which
outcomes should be aggregated.}

\item{L2.reg}{Geographic region. A character scalar containing the column
name of the geographic region in \code{survey} and \code{census} by which
geographic units are grouped (\code{L2.unit} must be nested within
\code{L2.reg}). Default is \code{NULL}.}

\item{kernel}{SVM kernel. A character-valued scalar specifying the kernel to
be used by SVM. The possible values are \code{linear}, \code{polynomial},
\code{radial}, and \code{sigmoid}. Default is \code{radial}.}

\item{loss.fun}{Loss function. A character-valued scalar indicating whether
prediction loss should be measured by the mean squared error (\code{MSE})
or the mean absolute error (\code{MAE}). Default is \code{MSE}.}

\item{loss.unit}{Loss function unit. A character-valued scalar indicating
whether performance loss should be evaluated at the level of individual
respondents (\code{individuals}), geographic units (\code{L2 units}) or at
both levels. Default is \code{c("individuals", "L2 units")}. With multiple
loss units, parameters are ranked for each loss unit and the loss unit with
the lowest rank sum is chosen. Ties are broken according to the order in
the search grid.}

\item{gamma}{SVM kernel parameter. A numeric vector whose values specify the
gamma parameter in the SVM kernel. This parameter is needed for all kernel
types except linear. Default is a sequence with minimum = 1e-5, maximum =
1e-1, and length = 20 that is equally spaced on the log-scale.}

\item{cost}{SVM cost parameter. A numeric vector whose values specify the
cost of constraints violation in SVM. Default is a sequence with minimum =
0.5, maximum = 10, and length = 5 that is equally spaced on the log-scale.}

\item{data}{Data for cross-validation. A \code{list} of \eqn{k}
\code{data.frames}, one for each fold to be used in \eqn{k}-fold
cross-validation.}

\item{verbose}{Verbose output. A logical argument indicating whether or not
verbose output should be printed. Default is \code{FALSE}.}

\item{cores}{The number of cores to be used. An integer indicating the number
of processor cores used for parallel computing. Default is 1.}
}
\value{
The support vector machine tuned parameters. A list.
}
\description{
\code{run_svm} is a wrapper function that applies the support vector machine
classifier to data provided by the user, evaluates prediction performance,
and chooses the best-performing model.
}
