% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_svm.R
\name{run_svm_mc}
\alias{run_svm_mc}
\title{SVM multicore tuning.}
\usage{
run_svm_mc(
  y,
  L1.x,
  L2.x,
  L2.eval.unit,
  L2.unit,
  L2.reg,
  form,
  loss.unit,
  loss.fun,
  data,
  cores,
  svm.grid,
  verbose
)
}
\arguments{
\item{L2.eval.unit}{Geographic unit. A character scalar containing the column
name of the geographic unit in \code{survey} and \code{census} at which
outcomes should be aggregated.}

\item{form}{The SVM model formula.}

\item{data}{Data for cross-validation. A \code{list} of \eqn{k}
\code{data.frames}, one for each fold to be used in \eqn{k}-fold
cross-validation.}

\item{svm.grid}{Search grid. A data.frame object where columns are parameters
and rows are search iterations.}
}
\value{
The cross-validation errors for all models. A list.
}
\description{
\code{run_svm_mc} is called from within \code{run_svm}. It tunes using
multiple cores.
}
