% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_classifier.R
\name{svm_classifier}
\alias{svm_classifier}
\title{SVM classifier}
\usage{
svm_classifier(
  method,
  form,
  data,
  kernel,
  error.fun,
  probability,
  svm.gamma,
  svm.cost,
  sampling = "cross",
  cross,
  verbose = c(TRUE, FALSE)
)
}
\arguments{
\item{method}{Function. A character string specifying the name of the
function to be tuned.}

\item{form}{Model formula. A two-sided linear formula describing
the model to be fit, with the outcome on the LHS and the covariates
separated by + operators on the RHS.}

\item{data}{Data. A data.frame containing the cross-validation data used to
train and evaluate the model.}

\item{kernel}{Kernel for SVM. A character string specifying the kernel to
be used for SVM. The possible types are linear, polynomial, radial, and
sigmoid. Default is radial.}

\item{error.fun}{function returning the error measure to be minimized. It
takes two arguments: a vector of true values and a vector of predicted values.
If NULL, the misclassification error is used for categorical predictions and
the mean squared error for numeric predictions.}

\item{probability}{Probability predictions. A logical argument indicating
whether the model should allow for probability predictions}

\item{svm.gamma}{Gamma parameter for SVM. This parameter is needed for all
kernels except linear.}

\item{svm.cost}{Cost parameter for SVM. This parameter specifies the cost of
constraints violation.}

\item{sampling}{Sampling scheme. A character string specifying the sampling
scheme to be used. Possible values are cross, fix, and boot. Default is
cross, which performs cross-validation.}

\item{cross}{The number cross-validation folds. A numeric scalar.}

\item{verbose}{Verbose output. A logical vector indicating whether or not
verbose output should be printed.}
}
\value{
The support vector machine model. An \code{\link[e1071]{tune}} object.
}
\description{
\code{svm_classifier} applies support vector machine classification to a
data set.
}
\examples{
\dontrun{
# Prepare data
survey_item <- dplyr::bind_rows(survey_item) \%>\%
  dplyr::mutate_at(.vars = y, as.factor)

# run svm classifier
models <- svm_classifier(
  method = "svm",
  form = YES ~ L1x1 + L1x2 + state + region + L2.x1,
  data = survey_item,
  kernel = "radial",
  error.fun = "MSE",
  probability = TRUE,
  gamma = c(0.3, 0.5, 0.55, 0.6, 0.65, 0.7, 0.8, 0.9, 1, 2, 3, 4),
  cost = c(1, 10),
  sampling = "cross",
  cross = 5,
  verbose = TRUE)
}
}
